pdf-chatbot/pyproject.toml
[project]
name = "pdf-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "huaan9426",email = "321210712@qq.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "langchain",
    "langchain-community",
    "openai",
    "dashscope",
    "sentence-transformers",
    "pypdf",
    "streamlit",
    "chromadb",
    "python-dotenv"
]

[tool.poetry]
packages = [{include = "pdf_chatbot", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
